+++ lib/toccatore/cli.rb
@@ -22,26 +22,15 @@ module Toccatore
     method_option :push_url, type: :string
     method_option :from_date, type: :string, default: (Time.now.to_date - 1.day).iso8601
     method_option :until_date, type: :string, default: Time.now.to_date.iso8601
+    method_option :query, type: :string
+    method_option :orcid, type: :string
+    method_option :doi, type: :string
     def orcid_update
+      options[:query] = "doi:#{options[:doi]}" if options[:doi].present?
+      options[:query] = "nameIdentifier:ORCID\\:#{options[:orcid]}" if options[:orcid].present?
+
       orcid_update = Toccatore::OrcidUpdate.new
-
-      data = orcid_update.get_data(options.merge(timeout: orcid_update.timeout, source_id: orcid_update.source_id))
-      data = orcid_update.parse_data(data, options.merge(source_id: orcid_update.source_id))
-
-      if data.empty?
-        puts "No works found for date range #{options[:from_date]} - #{options[:until_date]}."
-      else
-        data.each do |item|
-          response = orcid_update.push_item(item, options)
-          if response.body["data"].present?
-            doi = response.body.fetch("data", {}).fetch("attributes", {}).fetch("doi", nil)
-            orcid = response.body.fetch("data", {}).fetch("attributes", {}).fetch("orcid", nil)
-            puts "DOI #{doi} for ORCID ID #{orcid} pushed to Profiles service."
-          elsif response.body["errors"].present?
-            puts "An error occured: #{response.body['errors'].first['title']}"
-          end
-        end
-      end
+      orcid_update.queue_jobs(options)
     end
   end
 end