+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v.0.2 (February 2, 2017)
+
+[toccatore 0.2](https://github.com/datacite/toccatore/releases/tag/v.0.2) was released on February 2, 2017:
+
+* fixed bug of only sending 1,000 claims
+* enabled queries by DOI or ORCID ID
+
 ## v.0.1 (January 30, 2017)
 
 [toccatore 0.1](https://github.com/datacite/toccatore/releases/tag/v.0.1) was released on January 30, 2017: